Fix bug causing fish to not increase in weight
I wasn't passing fish as pointers to the obj_diff function so the
objective function wasn't being updated, causing the fish to not
increase in weight. I've rectified this by passing the fish as pointers
now. I've also made the print_fish function print a single fish of
choice for debugging purposes. To print all the fish in a table use
print_all_fish.