Add results of scheduling and parallel_for experiments
The results of each scheduling type are stored in their respective
directories. The experiments range from a chunk size of 10 to 10000. I
also added the results for varying fish number and step number in the
parallel_for implmentation. They're stored in parallel_diff_fish_step.
They follow the same naming rules as the sequential experiments.